chore: add unit tests for observers and pubsub client factory
